requirements.txt
absl-py==2.2.2
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
deepdiff==8.5.0
Deprecated==1.2.18
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.14.0
flake8==7.3.0
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
ignite==1.1.0
imageio==2.37.0
importlib_metadata==8.4.0
ipython==8.35.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
json-tricks==3.17.3
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.45.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmdet==3.0.0
mmengine==0.10.7
mmpose==1.3.2
model-index==0.1.11
monai==1.5.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.4.2
ninja==1.11.1.4
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
openTSNE==1.0.4
openxlab==0.1.2
ordered-set==4.1.0
orderly-set==5.5.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==3.0.0a2
shapely==2.1.0
simsimd==6.5.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.20.0a20250714
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.20
tomli==2.2.1
torch==2.6.0
torchmetrics==0.10.3
torchreid==0.2.5
torchvision==0.21.0
tqdm==4.65.2
traitlets==5.14.3
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2025.2
umap==0.1.1
umap-learn==0.5.9.post2
urllib3==1.26.20
vutils==1.1.8
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.43.0
zipp==3.23.0
